前端第三次修改

+++ webwork/src/main/java/service/SalaryService.java
@@ -21,7 +21,7 @@ public class SalaryService {
         // 总人数
         int count = salaryDao.count();
 
-        return n / count;
+        return (double) n / count;
     }
     public void uploadExcel(InputStream excelFileStream) throws IOException {
         // 读取Excel文件
